pyproject.toml
[project]
name = "nonebot-plugin-emojilike"
version = "0.2.1"
description = "nonebot2 名片赞，表情回应插件"
authors = [{ name = "fllesser", email = "fllessive@gmail.com" }]
readme = "README.md"
urls = { Repository = "https://github.com/fllesser/nonebot-plugin-emojilike" }
keywords = ["nonebot", "nonebot2", "emoji", "like"]
requires-python = ">=3.10"
dependencies = [
    "emoji>=2.14.1",
    "nonebot2>=2.4.2,<3.0.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot_plugin_localstore>=0.7.3,<1",
    "nonebot-plugin-apscheduler>=0.5.0,<1",
]

[tool.nonebot]
plugins = ["nonebot_plugin_emojilike"]

[dependency-groups]
dev = ["nonebot2[fastapi]>=2.4.2,<3.0.0", "ruff>=0.11.2,<1.0.0"]

test = [
    "nonebot2[fastapi]>=2.4.2,<3.0.0",
    "nonebot-adapter-onebot>=2.4.6,<3.0.0",
    "nonebug>=0.3.7,<1.0.0",
    "pytest-xdist>=3.6.1,<4.0.0",
    "pytest-asyncio>=0.26.0,<1.0.0",
]


[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
pythonpath = [".", "nonebot_plugin_template", "tests"]
addopts = [
    "-v",               # 详细输出
    "-s",               # 显示打印信息
    "--tb=short",       # 简短的错误回溯
    "-ra",              # 显示所有测试结果摘要
    "--strict-markers", # 严格标记模式
    # "--doctest-modules", # 运行文档测试
    "--import-mode=prepend", # 导入模式
]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
line-ending = "lf"

[tool.ruff.lint]
select = [
    "F",     # Pyflakes
    "W",     # pycodestyle warnings
    "E",     # pycodestyle errors
    "I",     # isort
    "UP",    # pyupgrade
    "ASYNC", # flake8-async
    "C4",    # flake8-comprehensions
    "T10",   # flake8-debugger
    "T20",   # flake8-print
    "PYI",   # flake8-pyi
    "PT",    # flake8-pytest-style
    "Q",     # flake8-quotes
    "TID",   # flake8-tidy-imports
    "RUF",   # Ruff-specific rules
]
ignore = [
    "E402",   # module-import-not-at-top-of-file
    "UP037",  # quoted-annotation
    "RUF001", # ambiguous-unicode-character-string
    "RUF002", # ambiguous-unicode-character-docstring
    "RUF003", # ambiguous-unicode-character-comment
    "W191",   # indentation contains tabs
    # "I001",   # isort: imports are incorrectly sorted
]


[tool.ruff.lint.isort]
force-sort-within-sections = true
known-first-party = ["tests/*"]
extra-standard-library = ["typing_extensions"]

[tool.ruff.lint.flake8-pytest-style]
fixture-parentheses = false
mark-parentheses = false

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true


[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
defineConstant = { PYDANTIC_V2 = true }
executionEnvironments = [
    { root = "./tests", extraPaths = [
        "./",
    ] },
    { root = "./" },
]
typeCheckingMode = "standard"
reportShadowedImports = false
disableBytesTypePromotions = true
